requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
llama-cpp-python==0.2.39
sse-starlette==2.0.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
